debug

+++ app/models/tokenTradeModel.go
@@ -196,13 +196,13 @@ func (this *Trade) GetFeeInfoList(page, rows, uid, opt int, date uint64, name st
 }
 
 //获取单日bibi交易手续费
-func (this *Trade)GetTodayFee()(uint64,error)  {
+func (this *Trade)GetTodayFee()(float64,error)  {
 	engine:=utils.Engine_token
 	sql:="SELECT fee FROM (SELECT FROM_UNIXTIME(deal_time,'%Y-%m-%d')days,SUM(fee_cny) fee FROM g_token.trade where states=2  ) t WHERE "
 	current:=time.Now().Format("2006-01-02 15:04:05")
 	current =fmt.Sprintf(" t.days='%s'", current[:10])
 	fee:=&struct {
-		Fee uint64
+		Fee float64
 	}{}
 
 	_,err:=engine.SQL(sql+current).Get(fee)
